Usegroup addition incapsulated and added to writer, not only namespace.
